+++ lib/cookie.rs
@@ -8,6 +8,8 @@ It can be disabled with the `cookies` feature.
 
 use std::fmt;
 
+use crate::common::decode_url;
+
 /// Represents a Cookie
 pub struct Cookie {
     /// Cookie Key
@@ -78,7 +80,10 @@ impl Cookie {
                 }
                 .trim_end_matches(';');
 
-                final_cookies.push(Cookie::new(name, value));
+                final_cookies.push(Cookie::new(
+                    decode_url(name.to_owned()),
+                    decode_url(value.to_owned()),
+                ));
             }
             return Some(final_cookies);
         }